refactor(session): extract AxesEditorMixin to _session_axes.py
Behavior-preserving move of _emit_axes, _set_axis, _set_offset_crosshair
into a mixin; Session now inherits from it.
